requirements.txt
addict==2.4.0
easydict==1.13
einops==0.8.2
flash_attn==2.8.3
huggingface_hub==0.36.2
matplotlib==3.10.8
numpy==2.4.4
pandas==3.0.2
Pillow==12.2.0
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
transformers==4.56.2
